Avoid load conflicts between node 0 and the target node.
In the past node 0 has served as the target of the many-to-one loading
for this test.  However, it is simultaneously serving as the host, if
you will, of the coforall-stmt across nodes.  Here, switch to using node
1 as the many-to-one target, so that node 0 doesn't have 2 jobs.  This
doesn't seem to have a very large impact on the performance results, but
it should only really affect the non-fast executeOn performance anyway,
and it still seems like the right thing to do.