Clean up measurement and reporting.
Improve the var names associated with performance measurement and
reporting, and also the phrases used to report the performance.
Change the .ml-keys and .ml-perf.graph files to match the new output
phrasing.

Note that if we adopt this change any performance .dat files containing
results for this test will have to be changed to match the new key
phrases also, or graph generation will fail.