Force 'this' to behave as parent type in phase one
InitNormalize will now store a VarSymbol 'mThisAsParent' which aliases
the data of 'this' but with the parent type and class ID.

As a result of this work, 'this' can be passed to functions and used to
call methods in phase one. Note that it is an error to call a method not
found in the parent type (or other ancestor types).

It is also now an error to cast 'this' in phase one.