Add overload for eye for CSDomain
for CSDomain. This simple fix adds that in, and includes
a simple test. Ideally, this should work for all
sparse domains, but I can't seem to figure out how
to get a more generic sparse domain.

As an added bonus, fix an issue where matPow would fail
for integer arrays.