Add a "registered" indicator to the memory region table entries.
With dynamic registration we allocate a region and and put it in the
table, then touch the memory (thus faultin it in) using an access
pattern we expect to produce the NUMA locality we desire, and only then
do we register it.  While we're touching the memory it is in the table
but is not yet registered.  Here, add to the memory region table entries
an indicator that says whether or not the corresponding regions have
actually been registered yet or not.  This indicator occupies the high
bit of the region length member, and there are new functions to encode
and decode the length and registration indicator to and from the length
values.

This is a step toward the future, to support catching a SIGBUS signal
that indicates insufficient hugepages precisely when we are faulting
them in without having reserved them first.

While here, also improve an overly generic variable name (mem_reg_addr
for the base address of the guaranteed-NIC-registered memory region, now
called gnr_addr), change chpl_comm_mem_reg_tell() accept NULL pointers
for values callers don't want, and get rid of a couple of unused vars.