Port Chapel to LLVM 6 (#8696)
Port Chapel to LLVM 6 (#8696)

[reviewed by @mppf ]

This change ports Chapel to LLVM 6. Some of the LLVM API has changed.
While there, ARM support is improved as well.
LLVM 6 is not made the default in this change. This only allows people to use LLVM 6 themselves. A forthcoming PR will make version 6 the default version of LLVM.

Passes full local testing.
Passes more of gasnet-fast than previous versions of LLVM.
Passes as much on Cray XC as previous versions of LLVM.

This is the first version of LLVM that fully supports the processors on the Cray XC50-AC ARM systems. Work remains to be done there.