Fix insertion of else block, and fix updated .good file
The else block wasn't getting properly inserted and so was thinking it didn't
have a parentExpr or parentSymbol, which caused problems for its resolution and
normalization.  Utilizing the insertion helper allowed us to fix that problem.

Also update the expected output for the test tracking this change, I had misread
the conditional.