Remove inCondStmt() and inOnInCondStmt() from InitNormalize
These functions were never used, and the tracking supporting them prevented us
from outputing the loop related error messages for field initialization in the
case where the initialization was wrapped in a conditional.  Remove them and
the tracking that supported them.

Update the expected output for the test that exposed this issue.