Futurize generic test tracking if statement error message
Since we now allow if statements without else branches by inserting
initialization, this test will now reach the same error impacting the other
version.  Also, it was setting a param field based on the value of a var field,
which is not allowed - update the test so it doesn't do that.