Temporarily update the expected error message for these tests
Due to the difference between how we handle initializer errors in loops and
errors in conditionals, and my removal of the tracking whether we are in a
conditional statement (which was not explicitly relied upon but was
unintentionally implicitly relied upon), these tests now generate the
loop-related error message instead of the conditional related one.

While this is mildly worse than before, it is not so terrible.  Removing the
tracking code prevented us from failing to report an error when we should, and
while the error we are reporting in this case is not the error that is closest
in scope, really we would prefer to report both errors and that is beyond the
scope of this PR.  Mike agrees that this is the right call for now.