Improve postInit support
This commit improves the functionality of postInit by enabling a number
of new patterns:
1) Compiler insertion of super.postInit
2) Handles absence of postInit between child and ancestor with postInit

The compiler will now generate postInit methods for types if an ancestor
type has a postInit. This allows all postInits in a heirarchy to be
called and works around a virtual dispatch bug.

Open question: what should we do with super.postInit calls in
conditionals?