Fix least common denominator CPU type for ARM-based Cray systems (#8825)
Fix least common denominator CPU type for ARM-based Cray systems (#8825)

[reviewed by @awallace-cray who will ensure the module build script stays in sync with this]

For a Cray XC system, we were always setting the least common denominator CPU architecture to "sandybridge". That is no longer always right, because of ARM systems. This change sets the LCD for ARM-based Cray XC systems to arm-thunderx2, which is currently the only ARM processor on Cray products.

This allows the ARM module to build without references to x86-64 architectures.