Add a snprintf()-like version of chpl_error().
This adds a new error function that is like chpl_error() in terms of
accepting a line number and filename index and formatting the message,
but instead of just a single message string it takes a printf()-style
format string and variable arguments, and it prints to a string instead
of to stdout.

This is a building block for a future use where we need to format an
error message in the standard chpl_error() way, but cannot print it
using the existing chpl_error() (and sibling) functions because we will
be executing in a signal handler.