Print usual out-of-mem message when unreserved hugepage fault-in fails.
We set the environment variable HUGETLB_NO_RESERVE=yes in order to get
hugepage NUMA locality matching the membind settings of the threads that
fault in those pages on first touch, rather than that of the thread that
allocates them.  (The first-touch is done in a forall-loop that iterates
in the same way we expect for later loops over the array, so we get good
NUMA affinity in those later loops.)  HUGETLB_NO_RESERVE=yes has the
side effect that allocating hugepages via for example get_huge_pages()
does not acquire the backing physical pages or even reserve them.
Instead, physical hugepages are acquired later, lazily, as they are
faulted in (touched), and if suitable memory cannot be found at that
time the kernel sends a SIGBUS signal.

In the past when this occurred the program just died due to the SIGBUS.
Here we arrange instead to catch the SIGBUS, print the usual "Out of
memory allocating <description>" message, and exit.  This involves
installing a handler for SIGBUS in the ugni comm layer iff we have the
HUGETLB_NO_RESERVE environment variable set, and then in the handler
seeing if the SIGBUS is due to a non-responding address (vs. one that is
misaligned, or a signal sent for some other reason), where the address
is within a hugepage-based memory region we have allocated but not yet
registered.  Any fault-in failure has to occur between these two events,
because before allocation we don't even have a mapping for the address,
so referencing it would give us a SIGSEGV instead of a SIGBUS, and after
registration the pages must exist because registration pins them.

There are supporting changes also.  mr_mregs_supplement[] is a vector of
memory descriptors, line numbers, and filename indexes which parallels
mem_regions.mregs[].  When we allocate a memory region we fill in the
corresponding mr_mregs_supplement[] element, which the SIGBUS handler
uses in the error message if the fault address is in that region.  The
interface for chpl_comm_regMemAlloc() is expanded to accept memory
description codes, line numbers, and filename indexes to put in this
supplemental vector, and the calls to chpl_comm_regMemAlloc() to
allocate arrays and heap extension blocks now pass the additional info.
There is also a new memory description code, for heap extension blocks.