Make testEnv call CHPL_TEST_UTIL_DIR's version of printchplenv
For some time now, `testEnv` has called the version of `printchplenv`
in CHPL_HOME.  This change makes it call the version in
CHPL_TEST_UTIL_DIR on the argument that it should be calling the
version that matches the scripts being used to run the test system
itself.  This caused a problem in generating release-over-release test
output this week because the printchplenv interface has changed so any
tests with skipif's were failing for versions of the release whose
printchplenv didn't support those flags.