Call chpl_exit_any() instead of exit().
I forgot that we've outlawed calling exit() from the runtime except in a
very few select cases.  For now, call chpl_exit_any() from the SIGBUS
handler and arrange to do none of the normal comm layer exit processing.
If this turns out not to work because of exit processing in other layers
we can arrange for an exceptional sys_exit() analogous to sys_malloc()
and its siblings.