Add lifetime checking (#8673)
This PR adds a lifetime checking component to the compiler which is off by 
default. The checker can be activated with the --lifetime-checking compiler 
flag or by using `pragma "safe"` on a function or module in which checking is
desired.

There are several language design questions that need to be answered before 
this checker can be enabled by default. These design questions are in many
issues but are linked to from #8373. The primary open question relevant to this
pass specifically is about how to override the default lifetime inference for
functions and methods - issue #8377. This PR uses pragmas for that purpose.
Most of the other open issues are about the evolution of `Owned` more
generally.

Lifetime checking allows for compile-time detection of many use-after-free 
errors. I have a prototype adding lifetime checking to the compiler in PR
#8338. The lifetime checker is concerned with finding potential use-after-free 
errors for references and borrows.

The lifetime checker infers the lifetime of each symbol (how long what it 
refers to will be allocated). Globals have infinite lifetime, but a local
simple value variable (e.g. an integer) has lifetime of the function block it 
is declared in. The lifetime of a symbol storing the result of a call is the
minimum of the actual lifetimes relevant to that call.

The checker emits errors for several cases:

* reachability of a symbol is greater than its lifetime 
* setting a symbol to another value with shorter lifetime
* returning a symbol with non-infinite and non-return lifetime 
  (e.g. reference arguments are marked with return lifetime)

The checker looks for misuses of `ref`/`const ref` in addition to borrows from 
Owned etc. The checker combines lifetimes for refs and lifetimes for borrows
into a single ScopeLifetime, but these can be conceptually be viewed as totally
unrelated - there could be 2 separate checking passes, one for refs, and one 
for borrows.

- [x] full local testing

Reviewed by @vasslitvinov - thanks!