Fix testEnv for use within paratest's directory gatherer
My previous change to testEnv assumed that $CHPL_TEST_UTIL_DIR was
set, which it is within start_test/sub_test scripts.  However, I
failed to notice that we also call 'testEnv' within paratest.server
to gather directory names, and it does not set the variable.  For
that reason, I changed the idiom to use CHPL_TEST_UTIL_DIR if it's
set, and to fall back to $CHPL_HOME/util if it is not.  This should
give us the benefit of my previous change without breaking other
traditional cases.