More readBytes() massaging
BHarsh pointed out that if the public interface only supported a
'generic ref' version that if someone passed in an actual c_ptr, then
it would get a ref to that cptr and read memory in from the wrong
place.  This tries to ammend that by making a public c_ptr version
as well.  That said, I still need to do more testing and cleanup.