Support dynamic dispatch on vararg methods
When a subclass overrides a vararg method defined by a parent class, dynamic
dispatch wasn't happening, so a call to the method through a variable that is
dynamically the subclass type but statically the parent class type would call
the parent version of the method.

When building up the list of methods for a parent class method dispatch set,
expand the varargs of the subclass methods to match the arguments the parent
method was vararg-expanded to.

This needed to look at calls to the parent method in order to expand the
subclass methods, so I moved most of the body of `compute_call_sites()` into
a new function `compute_fn_call_sites(FnSymbol*)`, to allow computing the call
sites for a single function.

Added the test from issue #6928.