Reinstall new SIGBUS handler after old one handles SIGBUS that isn't ours.
If a SIGBUS occurs and we determine it isn't for a fault-in failure on
one of our new allocations, we reinstall the old SIGBUS handler and
raise the signal again, so that it gets handled as it would have had our
handler never been installed.  But if execution continues after that (an
event I confess I hadn't thought of, SIGBUS being seemingly so fatal)
then it means the old handler dealt with the signal and we should
continue, with our new handler again present.  So, reinstall it iff the
call to raise() returns.