Fix segfault when class with constructor inherits from type with default init
It turns out the check for if the parent utilizes an initializer (either that it
defines or the compiler generated one) was only looking for if an explicit
initializer was in the hierarchy.  Instead, use the check to determine whether a
default initializer should be generated for the type and remove the
no-longer-used parentDefinesInitializer

This resolves two tests that were failing prior to this change.