Fix segfault when class with constructor inherits from type with default init
It turns out the check for if the parent utilizes an initializer (either that it
defines or the compiler generated one) was only looking for if an explicit
initializer was in the hierarchy.  Expand the check to also error when a default
initializer is present above the class with a constructor.