Switch revcomp-blc from readBytes() to read() of an array slice
In code reviewing my readBytes() clean-up with Michael, he pointed out
that binary I/O on an array slice felt like a much more natural way to
write this to him.  I think we must've shied away from it in the past
because our line-by-line reading options tanked performance wise due
to all the slicing.  But for this case, we're reading big enough
chunks that the slice performance seems to be in the noise.  This also
feels more Chapeltastic than readBytes() for this situation.