Fix --llvm-wide-opt after PR #8194
PR #8194 changed codegenFieldPtr in a way that
inadverdently broke --llvm-wide-opt.

In --llvm-wide-opt mode, wide pointers are represented
as address-space-100 pointers in the LLVM IR, so they
are still subject to GEP / load / etc.

codegenFieldPtr was using Type* castType for two purposes:
 1. to compute the GEP offset to use
 2. to cast the type to.

For --llvm-wide-opt mode, the first use needed to
use a "wide pointer" so it would code generate to
an address-space-100 pointer. But the second use
needed the non-wide pointer, since the wide pointer
doesn't actually contain the field.

This commit addresses this problem by moving some
fLLVMWideOpt conditionals later in the code. The
change to when ret.isLVPtr is set seems more correct
since base might have be modified after this condition
in the previous version.