Fix up arrArg*-{function,promote} futures
The -function versions now return an array
(vs a promoted expression) so work as expected.

The -promote versions are still futures but
array default intent has changed from ref since
these were filed. So I updated the .good files
to match my best guess of the original filer's intent
and added .bad files so that we can more easily
detect when these need to be updated.