Repair the runtime unit-testing side of error.h after #8836.
There is limited support for a "standalone" build of the runtime to
support certain kinds of unit testing.  When I did the work in #8836 I
threw together a standalone version of chpl_error_vs(), intending to
come back and correct it later before merging, but I never returned to
finish up.  I also never tested it, which would have demonstrated that I
hadn't done so.  Complete and test it now.