Fix handling of --compopts= in start_test
Fix how we preprocess and fixup --compopts= and --execopts= arguments to
start_test. We convert things like `--compopts --fast` to `--compopts '--fast'`
so that argparse doesn't interpret `--fast` as an option to start_test and we
later strip the leading and trailing characters to remove the quotes. However,
we were removing the leading and trailing characters even when we didn't add
quotes. for things like `--compopts=fast`. Correct that by only stripping the
leading and trailing characters if they were quotes.