call resolveSpecifiedReturnType only for candidates
Before this commit, resolveSpecifiedReturnType was called in
resolveSignature, which was called from checkResolveFormalsWhereClauses
during candidate filtering (and so was applied to all visible functions).
This commit changes resolveSpecifiedReturnType to only be called on
candidates that are potential matches - that is, those that have passed
argument type based filtering as well as where clause filtering.

Initially I tried to move the resolution of specified return types closer
to the inference of return types when the return type is not specified;
however a straightforward attempt in that direction breaks resolution of
recursive functions.

This change enables two example programs to work that did not work
before:
 * the example from issue #8131
 * the original example in issue #7836

Both of these examples are added as tests by this commit.