Make chpl tab completion pay attention to CHPL_DEVELOPER and --[no-]devel
Only complete developer options if CHPL_DEVELOPER is set or --devel was used
on the command line.

The flags --[no-]devel override the setting of CHPL_DEVELOPER.  If
CHPL_DEVELOPER is set but --no-devel is on the command line, do not complete
developer flags.  If CHPL_DEVELOPER is not set, but --devel is on the command
line, then do complete developer flags.