Upgrade from hwloc 1.11.9 to 1.11.10
This release includes a number of bugfixes and improvements.

A notable fix that's not mentioned in the release notes is that the check for
sysctl is skipped for linux. sysctl wasn't used for the linux backends, but the
check for it led to it being linked in, which caused linker warnings on some
aarch64 systems: https://github.com/open-mpi/hwloc/commit/58b867ac

For a full changelog see:
https://mail-archive.com/hwloc-announce@lists.open-mpi.org/msg00111.html