Limit resolution candidate searching for initializers based on type
Our existing resolution infrastructure finds many visible methods named
'init' that we the attempt to filter for 'good' candidates. The majority
of these initializers are not for the type that we're trying to
initialize.

We can get into trouble if we attempt to resolve the signature of a
candidate initializer because that candidate may be in the middle of
being initialized. For example, consider the default initializer for a
record storing an array of another record type:

```chpl
record R {
  var x : [1..3] OtherRec;
}
proc R.init(x = {var temp : [1..3] OtherRec; temp } ) { } // pseudocode
```

Resolution may look like this:
- resolve R.init()
- resolve R.init's formals
- resolve x's default value
- resolve [1..3] OtherRec
- resolve OtherRec.init
  - find visible 'init' methods
  - is 'R.init' a candidate?
    - resolve signature of R.init <---- RECURSED!

Instead of attempting to resolve the signature of every candidate
initializer, this commit adds a check to return early if the
initializer's 'this' formal and the corresponding actual are not
instantiated from the same type.