Fix a problem I introduced in compute_call_sites
When I pulled compute_fn_call_sites out of compute_call_sites, I cleared the
calledBy vector at the start of compute_fn_call_sites instead of clearing it
for all functions at once at the start of compute_call_sites.  In the presence
of virtual dispatch, a parent function will add to calledBy for all of the
child versions of the function.

Clearing it at the start of compute_fn_call_sites removes the calls for
virtual dispatch that were added by the parent version of the function.
Change the clearing back to being done all at once at the beginning of
compute_call_sites.