Add a future for the enum array default initializer bug, and a test that works
When converting initialize methods to use postinit, I discovered an error with a
record containing an enum array.  If the record utilized a default initializer
and default copy initializer, copies of the instance would not appropriately
copy the enum array field, leading to double deletes and accessing freed memory.
Explicit initializers do not seem to trigger this problem, oddly enough.

This commit adds a future for the error case and locks in the explicit
initializer case.  (Though my local machine was not as reliably reproducing the
error so we'll see if the "passing" version actually works.)