Update strassen test for reasonable error
The test had multiple errors and this PR changed the order
of reporting the errors. I fixed a new call to get the intended
behavior of this test. Then, the line number changed from the
return type expression to the return statement. Since the return
statement is involved, I think this is a reasonable line number.