Default Initializers: Parent formals before child formals
This commit updates default initializers such that the formals for
parent fields precede the formals for child fields. This matches the
existing behavior for constructors and matches the flow of
initialization now that parents are initialized before children.

To accomplish this, it was first necessary to remove some incorrect
error messages claiming that parent fields could not be accessed in
phase one. These were likely left over from a time before the new
initializer proposal was implemented.

Next it was necessary to update 'resolveMaybeSyncSingleField' to update
symbols in the field's type expression to refer to the instantiated
parent field instead of the fully generic field. This is accomplished by
adding a new field to AggregateType, 'genericFieldMap', which maps fully
generic fields from the root instantiation to concrete fields in another
instantiation. This map is built up as instantiations are created.