Add tests for various ways of specifying the library path
These tests test various combinations of CHPL_LIB_PATH and -L to make
sure that paths can be specified in both ways in an accumulative
manner, using ':'s as separators.  In order to create libraries that
can be linked into the code, I relied on C programs which may impact
the portability of this test w.r.t. different tool suites.