Unify handling of -L and -I flags
While making updates to the implementation of the -L flag in response
to Greg's comments, I looked ahead to how the -I flag was handled and
found it to be much simpler owing to its use of C++ (genius!).  Rather
than continue to carry the legacy C code along, I switched the -L
flag's handling to use C++-based techniques as well and then unified
the code to handle colon separation so that both would benefit from
it.

This involved separating -l flags from -L flags in the implementation
for the first time, but that doesn't seem like a big deal (and
actually seems cleaner).

TODO:
* add an environment variable to back -I
* change libflags testing to test various combinations of -I as well