Enable generic class heirarchies with initializers to use promotion
Some helper methods on AggregateType that create new instantiations were
incorrectly removing FLAG_GENERIC, which fooled some of the existing
promotion resolution machinery into not running. This commit updates
those helper methods to preserve FLAG_GENERIC if there are still generic
fields in the instantiation.

This commit also adds a call to resolvePromotionType for the case that
the parent is generic but the child is not.