Resolve destructors at end of resolution
Resolving initializer destructors at instantiation time is currently not
possible. If a field has a type expression that depends on a generic
field, it will not be resolved despite the type no longer being generic.
This is problematic if the depending field is used within the deinit.

A workaround is to resolve deinits at the end of resolution.