Be conservative about IPA (#9115)
Be conservative about IPA (#9115)

[reviewed by @ronawho ]

We don't always control which version of CCE is being used. This change conservatively backs off interprocedural analysis (IPA) optimizations slightly in case someone is using a version that has had difficulty with high levels of IPA.

Verified to work and to avoid under- or over-optimizing on x86-64 and ARM Cray systems.