Resolve point of instantiation bug for initializers
Constructors handled point of instantiation for virtually dispatched
methods by using the point of instantiation of the type constructor.
This commit does something similar for types with initializers by
storing the instantiation point of the initializer within the type once
that type is concrete.

For convenience this commit also wraps this functionality, along with
promotion resolution and name-fixing, into a helper function named
'resolveTypeWithInitializer'.