Do not RVF _array/_domain when passed by value
We were doing remote value forwarding (RVF) of wrapper records
even when they were passed into the on-function by value
i.e. '[const] in' intent. This is not useful, as the record fields
are passed to the remote locale as part of the record value even
without RVF.

This change restricts RVF of a wrapper record to the case
when it is passed by reference. This upholds the assertion that
the formal undergoing RVF has a ref intent.

While there, tightened up two switch statements to eliminate
the 'default' clause and to INT_FATAL if none of the cases match.
Commentary: bitwise operations on an IntentTag may result
in an illegal IntentTag, ex:
  INTENT_REF_MAYBE_CONST minus INTENT_FLAG_REF plus INTENT_FLAG_IN

Added a test of a peculiar condition that we apparently test
only on two tests in one configuration:

- An array and a domain are passed by forall 'in' intent.
- The parallel iterator - the range standalone, in these two tests -
  yields inside a task function that is coforall + on.

This resolves the failure introduced by #9080 in numa configuration
in these tests:

    parallel/forall/in-intents/both-arr-dom-const-const.chpl
    parallel/forall/in-intents/both-arr-dom-var-const.chpl