move lowering of ForallStmts to lowerIterators (#8785)
Lowering of ForallStmts at resolution - which is how it was
before this change - was unsatisfying.

This is because forall intents were implemented by extending the yield
expressions of the parallel iterator with task-local variables, which
are referenced as shadow variables in the loop body. This complicates
the compiler representation, for example by using tuples with references
that would not need to be handled otherwise.

Now ForallStmt nodes are preserved until lowerIterators and are
lowered by inlining the parallel iterator. Forall intents are
implemented directly by introducing VarSymbols for shadow variables.

### COMPILER

* The code that implemented forall intents for ForallStmt nodes
by extending yield tuples is now deleted.

* That does not include forall statements that are handled the old,
pre-ForallStmt way. Those forall statements still rely on old ways of
handling forall intents.

* With this PR, forall statements whose parallel iterators are
dynamically-dispatched are not implemented. This is no change
from before.

* With this PR, forall statements whose parallel iterators are
recursive are lowered using the same implementation as non-parallel
recursive iterators. This means that only [const]ref forall intents
are available for such forall loops. In this case ref forall intents
are implemented by replacing shadow variables in the loop body
with direct references to the outer variables.

* While thre, added TFI_TASK_PRIVATE to the enum and to switch statements
in support for the upcoming task-private variables.

* lowerForallStmts1()/lowerForallStmts2()/lowerForallIntents() are
distinct for historical reasons. Todo: see if they can be collapsed,
to simplify code organization.

* Can we now get rid of 'beforeLoweringForallStmts' ?

### TESTS

* Un-futurized the test:
        multilocale/vass/recursive-iterator-twice-multiloc

* Removed the following test as it is identical to the above under comm=none
  (all it does is remove on-clauses):
        functions/iterators/vass/recursive-iterator-twice-local

* Resolves #7159 i.e. compiler no longer crashes on this test:
        functions/iterators/vass/forall-in-standalone-iterator


Detailed development history: a96747c961..dc96be4c54

r: @mppf